Dont panic, this commit is just involving renaming the naming enforced by clang-tidy. These are just renaming changes with minimal changes built on top of the removal of global update

+++ src/atlas/core/window.cpp
@@ -47,7 +47,7 @@ namespace atlas {
     }
 
     float window::get_aspect_ratio() const {
-        return (float)get_width() / get_height();
+        return (float)get_width() / (float)get_height();
     }
 
     GLFWwindow* window::get_native_window() const {
